CourseList#new w/o parameters returns all courses
Calling #new with empty query values still returns an empty array. But we
sometimes need a way to get all of the courses, and since not *every*
course is associated with a Cluster, the only way to find some courses
is by searching by w/ specific info or addressing them specifically with
Verso::Course#new.